Update Classification_little_vgg.py
Error in The ImageDataGenerator is configured for grayscale images (color_mode='grayscale'), but the input shape of the first convolutional layer in the model does not match the expected shape for grayscale images. You need to add the number of channels (1 for grayscale) to the input shape.
so, Modify the input shape of the first convolutional layer to include the number of channels.