Update Classification_mobile_net.py
printMinMax function is defined but not called.

ModelCheckpoint is trying to monitor 'val_loss', but 'val_loss' might not be defined in your case. Change it to monitor 'val_accuracy' if you are using accuracy as a metric.

validation_datagen variable is used but not defined.

The linux_interaction() function is not defined.